refactor: 테스트 개선 및 폴더 구조 수정
