+++ src/main/java/com/dku/council/global/auth/jwt/AuthenticationTokenProvider.java
@@ -1,8 +1,5 @@
 package com.dku.council.global.auth.jwt;
 
-import com.dku.council.domain.user.User;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
 import org.springframework.security.core.Authentication;
 
 import javax.servlet.http.HttpServletRequest;
