+++ src/test/java/com/dku/council/global/auth/jwt/JwtAuthenticationTokenProviderTest.java
@@ -3,20 +3,14 @@ package com.dku.council.global.auth.jwt;
 import com.dku.council.domain.UserRole;
 import com.dku.council.domain.user.User;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class JwtAuthenticationTokenProviderTest {
 
-//    @Test
-//    void test1(){
-//        JwtAuthenticationTokenProvider jwtAuthenticationTokenProvider = new JwtAuthenticationTokenProvider();
-//        jwtAuthenticationTokenProvider.setAccessExpiration(3L);
-//        jwtAuthenticationTokenProvider.setRefreshExpiration(10L);
-//        jwtAuthenticationTokenProvider.setSecretKey("hello");
-//        AuthenticationToken issue = jwtAuthenticationTokenProvider.issue(new User(123L, UserRole.ADMIN));
-//        AuthenticationToken authenticationToken = jwtAuthenticationTokenProvider.reIssue(issue);
-//    }
+    @Test
+    void test1() {
+        JwtAuthenticationTokenProvider jwtAuthenticationTokenProvider = new JwtAuthenticationTokenProvider(3L, 10L, "hello");
+        AuthenticationToken issue = jwtAuthenticationTokenProvider.issue(new User(123L, UserRole.ADMIN));
+        AuthenticationToken authenticationToken = jwtAuthenticationTokenProvider.reIssue(issue);
+    }
 
 }
